# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderLineItemImage = {
  src?: string;
};

export interface OrderLineItem {
  id: number;
  name: string;
  quantity: number;
  price: string;
  total: string;
  sku?: string;
  image?: OrderLineItemImage;
}

export interface OrderBilling {
  first_name: string;
  last_name: string;
  email: string;
  phone?: string;
  address_1?: string;
  address_2?: string;
  city?: string;
  state?: string;
  postcode?: string;
  country?: string;
}

export interface OrderShipping {
  first_name: string;
  last_name: string;
  address_1?: string;
  address_2?: string;
  city?: string;
  state?: string;
  postcode?: string;
  country?: string;
}

export interface Order {
  id: number;
  number: string;
  status: string;
  date_created: string;
  date_modified?: string;
  total: string;
  subtotal?: string;
  total_tax?: string;
  shipping_total?: string;
  discount_total?: string;
  currency: string;
  currency_symbol?: string;
  payment_method?: string;
  payment_method_title?: string;
  customer_note?: string;
  billing: OrderBilling;
  shipping?: OrderShipping;
  line_items: OrderLineItem[];
}

export interface OrdersResponse {
  orders: Order[];
  total: number;
  total_pages: number;
}

export interface UpdateOrderRequest {
  status: string;
  customer_note?: string;
}

export interface OrderStats {
  total_orders: number;
  pending: number;
  processing: number;
  completed: number;
  cancelled: number;
  refunded: number;
  on_hold: number;
  total_revenue: string;
}

export type GetOrdersParams = {
  status?: string;
  search?: string;
  page?: number;
  per_page?: number;
};