# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WooCommerce Admin API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: WooCommerce order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: per_page
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/stats:
    get:
      operationId: getOrderStats
      tags: [orders]
      summary: Get order statistics
      responses:
        "200":
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OrderLineItem:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        quantity:
          type: number
        price:
          type: string
        total:
          type: string
        sku:
          type: string
        image:
          type: object
          properties:
            src:
              type: string
      required:
        - id
        - name
        - quantity
        - price
        - total
    OrderBilling:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        address_1:
          type: string
        address_2:
          type: string
        city:
          type: string
        state:
          type: string
        postcode:
          type: string
        country:
          type: string
      required:
        - first_name
        - last_name
        - email
    OrderShipping:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        address_1:
          type: string
        address_2:
          type: string
        city:
          type: string
        state:
          type: string
        postcode:
          type: string
        country:
          type: string
      required:
        - first_name
        - last_name
    Order:
      type: object
      properties:
        id:
          type: number
        number:
          type: string
        status:
          type: string
        date_created:
          type: string
        date_modified:
          type: string
        total:
          type: string
        subtotal:
          type: string
        total_tax:
          type: string
        shipping_total:
          type: string
        discount_total:
          type: string
        currency:
          type: string
        currency_symbol:
          type: string
        payment_method:
          type: string
        payment_method_title:
          type: string
        customer_note:
          type: string
        billing:
          $ref: "#/components/schemas/OrderBilling"
        shipping:
          $ref: "#/components/schemas/OrderShipping"
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
      required:
        - id
        - number
        - status
        - date_created
        - total
        - currency
        - billing
        - line_items
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: number
        total_pages:
          type: number
      required:
        - orders
        - total
        - total_pages
    UpdateOrderRequest:
      type: object
      properties:
        status:
          type: string
        customer_note:
          type: string
      required:
        - status
    OrderStats:
      type: object
      properties:
        total_orders:
          type: number
        pending:
          type: number
        processing:
          type: number
        completed:
          type: number
        cancelled:
          type: number
        refunded:
          type: number
        on_hold:
          type: number
        total_revenue:
          type: string
      required:
        - total_orders
        - pending
        - processing
        - completed
        - cancelled
        - refunded
        - on_hold
        - total_revenue