# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderBilling } from "./orderBilling";
import type { OrderLineItem } from "./orderLineItem";
import type { OrderShipping } from "./orderShipping";

export interface Order {
  id: number;
  number: string;
  status: string;
  date_created: string;
  date_modified?: string;
  total: string;
  subtotal?: string;
  total_tax?: string;
  shipping_total?: string;
  discount_total?: string;
  currency: string;
  currency_symbol?: string;
  payment_method?: string;
  payment_method_title?: string;
  customer_note?: string;
  billing: OrderBilling;
  shipping?: OrderShipping;
  line_items: OrderLineItem[];
}