# lib/api-zod/src/generated/types/orderLineItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineItemImage } from "./orderLineItemImage";

export interface OrderLineItem {
  id: number;
  name: string;
  quantity: number;
  price: string;
  total: string;
  sku?: string;
  image?: OrderLineItemImage;
}