# lib/api-zod/src/generated/types/orderShipping.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderShipping {
  first_name: string;
  last_name: string;
  address_1?: string;
  address_2?: string;
  city?: string;
  state?: string;
  postcode?: string;
  country?: string;
}