# lib/api-zod/src/generated/types/orderStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderStats {
  total_orders: number;
  pending: number;
  processing: number;
  completed: number;
  cancelled: number;
  refunded: number;
  on_hold: number;
  total_revenue: string;
}