# lib/api-zod/src/generated/types/ordersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export interface OrdersResponse {
  orders: Order[];
  total: number;
  total_pages: number;
}