# lib/api-zod/src/generated/types/updateOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrderRequest {
  status: string;
  customer_note?: string;
}